requirements3.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements3.txt requirements3.in
#
arrow==0.10.0             # via stravalib
beaker==1.8.1             # via pyramid-beaker
certifi==2017.4.17        # via requests
chardet==3.0.4            # via requests
click==6.7                # via pip-tools
cornice==2.4.0
couchdb==1.1              # via tractdb
first==2.0.1              # via pip-tools
forecastiopy==0.21
hupper==1.0               # via pyramid
idna==2.5                 # via requests
invoke==0.12.2
jinja2==2.9.6
markupsafe==1.0           # via jinja2
nose==1.3.7
olefile==0.44             # via pillow
pastedeploy==1.5.2        # via pyramid
pillow==4.1.1
pip-tools==1.9.0
pyramid-beaker==0.8
pyramid==1.8.4
python-dateutil==2.6.0    # via arrow
pytz==2017.2              # via stravalib
pyyaml==3.12
repoze.lru==0.6           # via pyramid
requests==2.17.3
simplejson==3.10.0        # via cornice
six==1.10.0               # via pip-tools, python-dateutil, stravalib
stravalib==0.6.6
tractdb==0.1.14
translationstring==1.3    # via pyramid
units==0.7                # via stravalib
urllib3==1.21.1           # via requests
venusian==1.1.0           # via pyramid
waitress==1.0.2
webob==1.7.2              # via pyramid
zope.deprecation==4.2.0   # via pyramid
zope.interface==4.4.1     # via pyramid
